+++ cranelift/filetests/src/lib.rs
@@ -30,7 +30,6 @@ mod test_run;
 mod test_safepoint;
 mod test_unwind;
 mod test_verifier;
-pub mod test_wasm;
 
 /// Main entry point for `clif-util test`.
 ///
